third_party/WebKit/LayoutTests/inspector/device-emulation/device-emulation-scale-only-expected.txt

Tests that overriding only device scale factor does not affect any other value.

Overriding device scale factor: 1
Got device scale factor: 1
Overriding device scale factor: 2
Got device scale factor: 2
Overriding device scale factor: 3
Got device scale factor: 3
Overriding device scale factor: 0
